BUG FIX (Performance): Correct exponential time behaviour of TreeFree.optimize
The problem is that we update the sets of defined variables after trying to
append a free statement.  This will update the information in all statements
that follow one we tried to append to.  We correct it by checking whether
the information changed when computing the defined variable sets and do not
update the following statements, because the information will not change.


git-svn-id: https://svn.ifi.uio.no/repos/projects/credo/trunk/creoltools@2050 29972ad9-4e1c-0410-bd41-baaa3728e1af